pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyMapz"
description = "A Python Maze Traversal Visualizer for applying search algorithms."
version = "1.0"
requires-python = ">=3.11"
license = {text = "MIT"}
readme = "README.md"
authors =  [
    {name = "António Raimundo", email = "aslro@iscte-iul.pt"},
    {name = "Ricardo Pereira", email = "rtcpa@iscte-iul.pt"},
]

dependencies = [
    "networkx==3.4.2",
    "pygame-ce==2.5.5",
    "numpy==2.1.3",
    "matplotlib==3.9.2",
    "scipy==1.14.1",
    "pygame-gui==0.6.13",
    "geopandas==1.1.1",
]

[project.scripts]
pymapz = "main:main"
